### Summary
This PR adds implementations of the qpy module `dump` and `load`
functions with Rust, making it possible to have a full QPY flow directly
in Rust without Python.

### Details and comments
This PR enables loading and dumping QPY files directly from rust,
exposing `dump` and `load` functions to Python. Currently we do not
support dumping to QPY version < 17 and loading only for QPY version
>=13, so the default path remains through the Python `qpy` module's
`dump` and `load` functions which call the rust functions whenever
possible.

A new struct, `QPYFileHeader` was added to parse the remaining data
parsed by Python until now. Additionally, the offset table for the
circuits in the file is computed correctly when dumping, and the raw
bytes for the circuits are split correctly when loading; adding
rust-based parallelized code is one step away.

Since we still don't have a pure rust `QuantumCircuit`, the rust
functions `dump_qpy` and `load_qpy` still have Python remnants in their
I/O. In particular, `load_qpy` returns python circuits (as `Py<PyAny>`)
and `dump_qpy` accepts the `QuantumCircuitData` struct which is used as
a transition layer from python's `QuantumCircuit`.

+++ crates/qpy/src/circuit_writer.rs
@@ -1206,7 +1206,7 @@ pub(crate) fn pack_circuit(
     circuit: &mut QuantumCircuitData,
     metadata_serializer: Option<&Bound<PyAny>>,
     _use_symengine: bool,
-    version: u32,
+    version: u8,
     annotation_factories: &Bound<PyDict>,
 ) -> Result<formats::QPYCircuit, QpyError> {
     let annotation_handler = AnnotationHandler::new(annotation_factories);
@@ -1259,7 +1259,7 @@ pub(crate) fn py_write_circuit(
     circuit: &Bound<PyAny>,
     metadata_serializer: &Bound<PyAny>,
     use_symengine: bool,
-    version: u32,
+    version: u8,
     annotation_factories: &Bound<PyDict>,
 ) -> PyResult<usize> {
     let packed_circuit = pack_circuit(
